# docs/openapi/stigmem.json
{
  "components": {
    "schemas": {
...
      "Constraint": {
        "description": "Hard limit on an intent (Spec-X8-Intent-Envelope).",
        "properties": {
          "kind": {
            "minLength": 1,
            "title": "Kind",
            "type": "string"
          },
          "limit": {
            "$ref": "#/components/schemas/FactValue"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "required": [
          "kind",
          "limit"
        ],
        "title": "Constraint",
        "type": "object"
      },
      "DeferenceRule": {
        "description": "When to defer to another entity before proceeding (Spec-X8-Intent-Envelope).",
        "properties": {
          "condition": {
            "minLength": 1,
            "title": "Condition",
            "type": "string"
          },
          "defer_to": {
            "minLength": 1,
            "title": "Defer To",
            "type": "string"
          },
          "timeout_s": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout S"
          }
        },
        "required": [
          "condition",
          "defer_to"
        ],
        "title": "DeferenceRule",
        "type": "object"
      },
      "EscalationPolicy": {
        "description": "Who to notify and how when escalation is required (Spec-X8-Intent-Envelope).",
        "properties": {
          "channel": {
            "default": "stigmem",
            "title": "Channel",
            "type": "string"
          },
          "escalate_to": {
            "minLength": 1,
            "title": "Escalate To",
            "type": "string"
          },
          "include_context": {
            "default": true,
            "title": "Include Context",
            "type": "boolean"
          },
          "priority": {
            "default": "medium",
            "title": "Priority",
            "type": "string"
          }
        },
        "required": [
          "escalate_to"
        ],
        "title": "EscalationPolicy",
        "type": "object"
      },
...
      "FactValue": {
        "properties": {
          "type": {
            "title": "Type",
            "type": "string"
          },
          "v": {
            "title": "V"
          }
        },
        "required": [
          "type"
        ],
        "title": "FactValue",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HandoffArtifact": {
        "description": "Named artifact reference passed along with a handoff (Spec-X8-Intent-Envelope).",
        "properties": {
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "ref": {
            "description": "URI of the artifact fact",
            "minLength": 1,
            "title": "Ref",
            "type": "string"
          }
        },
        "required": [
          "name",
          "ref"
        ],
        "title": "HandoffArtifact",
        "type": "object"
      },
      "HandoffPayload": {
        "description": "Structured context transfer for agent handoffs (Spec-X8-Intent-Envelope).\n\nfact_refs MUST include any facts the receiver needs to reconstitute context.",
        "properties": {
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/HandoffArtifact"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "continuation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Continuation"
          },
          "fact_refs": {
            "items": {
              "type": "string"
            },
            "title": "Fact Refs",
            "type": "array"
          },
          "summary": {
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "summary"
        ],
        "title": "HandoffPayload",
        "type": "object"
      },
      "IntentEnvelopeRecord": {
        "description": "Response for POST /v1/intents and GET /v1/intents/:id (Spec-X8-Intent-Envelope).",
        "properties": {
          "constraint": {
            "items": {
              "$ref": "#/components/schemas/Constraint"
            },
            "title": "Constraint",
            "type": "array"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "deference": {
            "items": {
              "$ref": "#/components/schemas/DeferenceRule"
            },
            "title": "Deference",
            "type": "array"
          },
          "escalation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EscalationPolicy"
              },
              {
                "type": "null"
              }
            ]
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "fact_ids": {
            "description": "IDs of all facts written into the fabric for this envelope",
            "items": {
              "type": "string"
            },
            "title": "Fact Ids",
            "type": "array"
          },
          "from": {
            "description": "Sender entity URI",
            "title": "From",
            "type": "string"
          },
          "goal": {
            "title": "Goal",
            "type": "string"
          },
          "handoff": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HandoffPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "preference": {
            "items": {
              "$ref": "#/components/schemas/Preference"
            },
            "title": "Preference",
            "type": "array"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "to": {
            "items": {
              "type": "string"
            },
            "title": "To",
            "type": "array"
          }
        },
        "required": [
          "id",
          "from",
          "to",
          "goal",
          "scope",
          "created_at",
          "expires_at",
          "constraint",
          "preference",
          "deference",
          "escalation",
          "handoff",
          "fact_ids"
        ],
        "title": "IntentEnvelopeRecord",
        "type": "object"
      },
      "IntentEnvelopeRequest": {
        "description": "POST /v1/intents request body (Spec-X8-Intent-Envelope).\n\nThe ``from`` field is aliased as ``from_uri`` in Python to avoid the keyword clash.\nClients MUST send the JSON key ``\"from\"``.",
        "properties": {
          "constraint": {
            "items": {
              "$ref": "#/components/schemas/Constraint"
            },
            "title": "Constraint",
            "type": "array"
          },
          "deference": {
            "items": {
              "$ref": "#/components/schemas/DeferenceRule"
            },
            "title": "Deference",
            "type": "array"
          },
          "escalation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EscalationPolicy"
              },
              {
                "type": "null"
              }
            ]
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO 8601 UTC expiry; null = non-expiring",
            "title": "Expires At"
          },
          "from": {
            "description": "Sender entity URI",
            "minLength": 1,
            "title": "From",
            "type": "string"
          },
          "goal": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Goal",
            "type": "string"
          },
          "handoff": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HandoffPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional client-generated intent ID (UUID). Auto-assigned when omitted.",
            "title": "Id"
          },
          "preference": {
            "items": {
              "$ref": "#/components/schemas/Preference"
            },
            "title": "Preference",
            "type": "array"
          },
          "scope": {
            "default": "company",
            "title": "Scope",
            "type": "string"
          },
          "to": {
            "description": "Target entity URIs (at least one required)",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "To",
            "type": "array"
          }
        },
        "required": [
          "from",
          "to",
          "goal"
        ],
        "title": "IntentEnvelopeRequest",
        "type": "object"
      },
...
      "Preference": {
        "description": "Soft preference on an intent (Spec-X8-Intent-Envelope).",
        "properties": {
          "kind": {
            "minLength": 1,
            "title": "Kind",
            "type": "string"
          },
          "value": {
            "$ref": "#/components/schemas/FactValue"
          },
          "weight": {
            "default": 1.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "kind",
          "value"
        ],
        "title": "Preference",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Reference node implementing the Stigmem v0.9.0a1 HTTP API \u2014 facts, federation, gardens, recall, subscriptions, audit, identity, content-addressed fact IDs. Cross-cutting features (tombstones, time-travel, multi-tenant) are opt-in plugins.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title": "Stigmem Reference Node",
    "version": "0.9.0a1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/intents": {
      "post": {
        "description": "Submit an IntentEnvelope (Spec-X8-Intent-Envelope).\n\nValidates the envelope, normalizes URIs, decomposes it into atomic facts,\nand returns a receipt with the generated intent ID and all written fact IDs.\nIdempotent when the caller supplies a stable ``id``: if that intent_id already\nhas a ``intent:goal`` fact in the fabric, returns 409 Conflict.",
        "operationId": "submit_intent_v1_intents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntentEnvelopeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntentEnvelopeRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit Intent",
        "tags": [
          "intents"
        ]
      }
    },
    "/v1/intents/{intent_id}": {
      "get": {
        "description": "Retrieve an IntentEnvelope by ID, reconstructed from its reified facts.\n\nCovered by Spec-X8-Intent-Envelope.",
        "operationId": "get_intent_v1_intents__intent_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "intent_id",
            "required": true,
            "schema": {
              "title": "Intent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntentEnvelopeRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Intent",
        "tags": [
          "intents"
        ]
      }
    },
...
  }
}